some local changes

+++ data/dialog-success.sh
@@ -1,6 +1,11 @@
 #! /bin/bash
 
-if zenity --question --text ""$1" has been processed successfully. Would you like to reboot for changes to take effect?"
+if echo $1 | grep -i mesa
 then
-	systemctl reboot
+    true
+else
+    if zenity --question --text ""$1" has been processed successfully. Would you like to reboot for changes to take effect?"
+    then
+	    systemctl reboot
+    fi
 fi
